% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectionDAG.R
\name{create_inspection_dag}
\alias{create_inspection_dag}
\title{Create an Inspection DAG}
\usage{
create_inspection_dag(csv_path)
}
\arguments{
\item{csv_path}{A character string representing the file path to the tasks configuration CSV.}
}
\value{
An object of class \code{InspectionDAG}.
}
\description{
Reads a CSV configuration matrix, validates the workflow dependencies,
and constructs an \code{InspectionDAG} object.
}
