% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectionSimulation.R
\name{create_simulation_env}
\alias{create_simulation_env}
\title{Create a Simulation Environment}
\usage{
create_simulation_env(dag_csv, floorplan_csv, yaml_path, mix_weights)
}
\arguments{
\item{dag_csv}{Path to the tasks configuration CSV file.}

\item{floorplan_csv}{Path to the floorplans configuration CSV file.}

\item{yaml_path}{Path to the scenario configuration YAML file.}

\item{mix_weights}{A named numeric vector representing the distribution of unit types.}
}
\value{
An object of class \code{InspectionSimulation}.
}
\description{
Validates and constructs an \code{InspectionSimulation} object from raw configuration files.
}
