% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationEngine.R
\name{generate_building}
\alias{generate_building}
\title{Generate Building Demographics}
\usage{
generate_building(sim_env, total_units)
}
\arguments{
\item{sim_env}{An \code{InspectionSimulation} object.}

\item{total_units}{Integer. The total number of units in the building.}
}
\value{
A character vector of length \code{total_units} containing the assigned archetypes.
}
\description{
Spawns a physical building profile by sampling unit archetypes based on the
specified building mix probabilities.
}
