% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectionDAG.R
\docType{class}
\name{InspectionDAG-class}
\alias{InspectionDAG-class}
\title{InspectionDAG Class}
\description{
An S4 class that represents the mathematical directed acyclic graph (DAG)
of inspection tasks and their dependencies.
}
\section{Slots}{

\describe{
\item{\code{tasks_config}}{A \code{data.frame} containing task times, dependencies, and decay probabilities.}

\item{\code{adj_list}}{A \code{list} providing a fast-lookup prerequisite adjacency list.}

\item{\code{graph}}{An \code{igraph} object representing the network of task dependencies.}
}}

