% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectionSimulation.R
\docType{class}
\name{InspectionSimulation-class}
\alias{InspectionSimulation-class}
\title{InspectionSimulation Class}
\description{
An S4 class representing the complete environment for an NSPIRE Monte Carlo simulation.
}
\section{Slots}{

\describe{
\item{\code{dag}}{An object of class \code{InspectionDAG} representing the task dependencies.}

\item{\code{floorplans}}{A \code{data.frame} containing the physical item counts per unit archetype.}

\item{\code{scenario}}{A \code{list} containing the parsed YAML operational strategy parameters.}

\item{\code{building_mix}}{A \code{numeric} vector of normalized probabilities for unit archetypes.}
}}

