% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationEngine.R
\name{rpert}
\alias{rpert}
\title{Beta-PERT Random Number Generator}
\usage{
rpert(n, a, m, b)
}
\arguments{
\item{n}{Integer. Number of observations.}

\item{a}{Numeric. Minimum (optimistic) value.}

\item{m}{Numeric. Most likely (mode) value.}

\item{b}{Numeric. Maximum (pessimistic) value.}
}
\value{
A numeric vector of length n.
}
\description{
Generates a random value from a PERT distribution using the minimum, most likely,
and maximum parameters. Perfect for modeling task durations.
}
