% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectionSimulation.R
\name{show,InspectionSimulation-method}
\alias{show,InspectionSimulation-method}
\title{Show method for InspectionSimulation}
\usage{
\S4method{show}{InspectionSimulation}(object)
}
\arguments{
\item{object}{An \code{InspectionSimulation} object.}
}
\description{
Prints a clean dashboard summary of the simulation environment parameters.
}
